Refactor TornadoPartialLoopUnroll to common package
Deleted TornadoPartialLoopUnroll from OpenCL and SPIRV drivers, and created a similar file under common driver. Reduced code duplication by relocating the common implementations into a shared module (drivers common). This improves maintainability and code reuse. Also updated OCLMidTier and SPIRVMidTier to include the common TornadoPartialLoopUnroll.
